# Optimize benchmarks; permit serialization