Prior to these changes:

```
BenchmarkDeducedSimple-12                                   3000            472665 ns/op          139180 B/op       1810 allocs/op
BenchmarkDeducedNested-12                                   1000           1370363 ns/op          401776 B/op       5146 allocs/op
BenchmarkDeducedNestedAcrossVersion-12                      1000           1435743 ns/op          434426 B/op       5527 allocs/op
BenchmarkLeafConflictAcrossVersion-12                       3000            492026 ns/op          147113 B/op       1848 allocs/op
BenchmarkMultipleApplierRecursiveRealConversion-12          1000           2585529 ns/op          805147 B/op       8457 allocs/op
```

After these changes:

```
BenchmarkDeducedSimple-12                                   5000            432763 ns/op          127173 B/op       1703 allocs/op
BenchmarkDeducedNested-12                                   1000           1298198 ns/op          356460 B/op       4678 allocs/op
BenchmarkDeducedNestedAcrossVersion-12                      1000           1359040 ns/op          383408 B/op       5022 allocs/op
BenchmarkLeafConflictAcrossVersion-12                       3000            437574 ns/op          134185 B/op       1783 allocs/op
BenchmarkMultipleApplierRecursiveRealConversion-12           500           2382065 ns/op          732700 B/op       7978 allocs/op

```

Mostly I was looking at allocations but this is a modest improvement for everything.

This will permit a straightforward serialization. The current map with string key form can't be deserialized in a single step.