@lavalamp 2019-07-17
Antoine is OOO. I think this should be ready to go. I haven't tried this on the main repo but this is definitely not worse than what we have now unless the benchmarks are completely unrepresentative.

/assign @jennybuckley @kwiesmueller