@apelisse 2019-07-17
Benchmarks in this repo use significantly smaller sets than real-life sets. I doubt this is going to be equally good with real-life size objects :-/