@lavalamp 2019-07-17
Benchmark comparison:

```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkDeducedSimple-12                              472665        432763        -8.44%
BenchmarkDeducedNested-12                              1370363       1298198       -5.27%
BenchmarkDeducedNestedAcrossVersion-12                 1435743       1359040       -5.34%
BenchmarkLeafConflictAcrossVersion-12                  492026        437574        -11.07%
BenchmarkMultipleApplierRecursiveRealConversion-12     2585529       2382065       -7.87%

benchmark                                              old allocs     new allocs     delta
BenchmarkDeducedSimple-12                              1810           1703           -5.91%
BenchmarkDeducedNested-12                              5146           4678           -9.09%
BenchmarkDeducedNestedAcrossVersion-12                 5527           5022           -9.14%
BenchmarkLeafConflictAcrossVersion-12                  1848           1783           -3.52%
BenchmarkMultipleApplierRecursiveRealConversion-12     8457           7978           -5.66%

benchmark                                              old bytes     new bytes     delta
BenchmarkDeducedSimple-12                              139180        127173        -8.63%
BenchmarkDeducedNested-12                              401776        356460        -11.28%
BenchmarkDeducedNestedAcrossVersion-12                 434426        383408        -11.74%
BenchmarkLeafConflictAcrossVersion-12                  147113        134185        -8.79%
BenchmarkMultipleApplierRecursiveRealConversion-12     805147        732700        -9.00%
```