@lavalamp 2019-07-17
The benchmark code here is running a bunch of test assertions, diluting the signal here. After making it not do that, the results are not changed much, maybe slightly worse. Time seems pretty noisy, but I'm looking at allocations.

I started to make the items in the benchmarks bigger but that is a harder task than expected, plus I seem to have hit a bug where field names are not handled right, e.g. letter followed by number has the number stripped.

```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkDeducedSimple-12                              453044        458392        +1.18%
BenchmarkDeducedNested-12                              1382705       1266794       -8.38%
BenchmarkDeducedNestedAcrossVersion-12                 1501691       1388808       -7.52%
BenchmarkLeafConflictAcrossVersion-12                  465940        454865        -2.38%
BenchmarkMultipleApplierRecursiveRealConversion-12     2498657       2328747       -6.80%

benchmark                                              old allocs     new allocs     delta
BenchmarkDeducedSimple-12                              1810           1703           -5.91%
BenchmarkDeducedNested-12                              5146           4678           -9.09%
BenchmarkDeducedNestedAcrossVersion-12                 5527           5022           -9.14%
BenchmarkLeafConflictAcrossVersion-12                  1848           1783           -3.52%
BenchmarkMultipleApplierRecursiveRealConversion-12     8149           7694           -5.58%

benchmark                                              old bytes     new bytes     delta
BenchmarkDeducedSimple-12                              139179        127172        -8.63%
BenchmarkDeducedNested-12                              401784        356427        -11.29%
BenchmarkDeducedNestedAcrossVersion-12                 434410        383419        -11.74%
BenchmarkLeafConflictAcrossVersion-12                  147112        134184        -8.79%
BenchmarkMultipleApplierRecursiveRealConversion-12     784542        711478        -9.31%
```