@lavalamp 2019-07-18
I've added a microbenchmark for set.

```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkFieldSet/insert-10-12             11498         6663          -42.05%
BenchmarkFieldSet/has-10-12                910           748           -17.80%
BenchmarkFieldSet/union-10-12              4195          4132          -1.50%
BenchmarkFieldSet/intersection-10-12       1903          2790          +46.61%
BenchmarkFieldSet/difference-10-12         1893          3152          +66.51%
BenchmarkFieldSet/insert-20-12             58147         26062         -55.18%
BenchmarkFieldSet/has-20-12                1710          1037          -39.36%
BenchmarkFieldSet/union-20-12              41174         15651         -61.99%
BenchmarkFieldSet/intersection-20-12       8791          7893          -10.21%
BenchmarkFieldSet/difference-20-12         28758         12580         -56.26%
BenchmarkFieldSet/insert-50-12             186330        89396         -52.02%
BenchmarkFieldSet/has-50-12                1978          1354          -31.55%
BenchmarkFieldSet/union-50-12              131180        40162         -69.38%
BenchmarkFieldSet/intersection-50-12       21577         17659         -18.16%
BenchmarkFieldSet/difference-50-12         93593         30367         -67.55%
BenchmarkFieldSet/insert-100-12            676984        278113        -58.92%
BenchmarkFieldSet/has-100-12               2794          2016          -27.85%
BenchmarkFieldSet/union-100-12             262925        94525         -64.05%
BenchmarkFieldSet/intersection-100-12      31648         43703         +38.09%
BenchmarkFieldSet/difference-100-12        162427        74614         -54.06%
BenchmarkFieldSet/insert-500-12            3433305       1595039       -53.54%
BenchmarkFieldSet/has-500-12               3517          2639          -24.96%
BenchmarkFieldSet/union-500-12             1247269       392640        -68.52%
BenchmarkFieldSet/intersection-500-12      214700        183495        -14.53%
BenchmarkFieldSet/difference-500-12        873070        294497        -66.27%
BenchmarkFieldSet/insert-1000-12           8943251       3711783       -58.50%
BenchmarkFieldSet/has-1000-12              4074          2888          -29.11%
BenchmarkFieldSet/union-1000-12            2432131       714658        -70.62%
BenchmarkFieldSet/intersection-1000-12     461434        342822        -25.71%
BenchmarkFieldSet/difference-1000-12       1699348       586396        -65.49%

benchmark                                  old allocs     new allocs     delta
BenchmarkFieldSet/insert-10-12             50             31             -38.00%
BenchmarkFieldSet/has-10-12                4              5              +25.00%
BenchmarkFieldSet/union-10-12              6              21             +250.00%
BenchmarkFieldSet/intersection-10-12       5              19             +280.00%
BenchmarkFieldSet/difference-10-12         4              19             +375.00%
BenchmarkFieldSet/insert-20-12             231            134            -41.99%
BenchmarkFieldSet/has-20-12                8              6              -25.00%
BenchmarkFieldSet/union-20-12              127            85             -33.07%
BenchmarkFieldSet/intersection-20-12       45             61             +35.56%
BenchmarkFieldSet/difference-20-12         102            74             -27.45%
BenchmarkFieldSet/insert-50-12             743            446            -39.97%
BenchmarkFieldSet/has-50-12                9              7              -22.22%
BenchmarkFieldSet/union-50-12              416            196            -52.88%
BenchmarkFieldSet/intersection-50-12       93             126            +35.48%
BenchmarkFieldSet/difference-50-12         302            163            -46.03%
BenchmarkFieldSet/insert-100-12            2462           1338           -45.65%
BenchmarkFieldSet/has-100-12               11             9              -18.18%
BenchmarkFieldSet/union-100-12             1241           435            -64.95%
BenchmarkFieldSet/intersection-100-12      205            305            +48.78%
BenchmarkFieldSet/difference-100-12        818            388            -52.57%
BenchmarkFieldSet/insert-500-12            13424          8013           -40.31%
BenchmarkFieldSet/has-500-12               13             12             -7.69%
BenchmarkFieldSet/union-500-12             6027           2341           -61.16%
BenchmarkFieldSet/intersection-500-12      1215           1604           +32.02%
BenchmarkFieldSet/difference-500-12        4240           1998           -52.88%
BenchmarkFieldSet/insert-1000-12           29646          17782          -40.02%
BenchmarkFieldSet/has-1000-12              14             13             -7.14%
BenchmarkFieldSet/union-1000-12            11738          4413           -62.40%
BenchmarkFieldSet/intersection-1000-12     2198           2848           +29.57%
BenchmarkFieldSet/difference-1000-12       8040           3756           -53.28%

```

As you can see, it's a massive improvement with a few exceptions, mostly intersection.