@lavalamp 2019-07-18
I tweaked the benchmarks a bit. I did three runs of old and new, which were all very consistent with each other. Here's a random old / new comparison:

```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkFieldSet/insert-20-12             54252         24868         -54.16%
BenchmarkFieldSet/has-20-12                1597          1306          -18.22%
BenchmarkFieldSet/union-20-12              36040         17464         -51.54%
BenchmarkFieldSet/intersection-20-12       7631          8752          +14.69%
BenchmarkFieldSet/difference-20-12         25865         13741         -46.87%
BenchmarkFieldSet/insert-50-12             149605        90214         -39.70%
BenchmarkFieldSet/has-50-12                1776          1438          -19.03%
BenchmarkFieldSet/union-50-12              95960         37428         -61.00%
BenchmarkFieldSet/intersection-50-12       16027         15331         -4.34%
BenchmarkFieldSet/difference-50-12         67187         28200         -58.03%
BenchmarkFieldSet/insert-100-12            498853        233598        -53.17%
BenchmarkFieldSet/has-100-12               2357          1864          -20.92%
BenchmarkFieldSet/union-100-12             246561        73341         -70.25%
BenchmarkFieldSet/intersection-100-12      34281         33725         -1.62%
BenchmarkFieldSet/difference-100-12        170985        63803         -62.69%
BenchmarkFieldSet/insert-500-12            2797068       1407020       -49.70%
BenchmarkFieldSet/has-500-12               2607          2375          -8.90%
BenchmarkFieldSet/union-500-12             1095431       383595        -64.98%
BenchmarkFieldSet/intersection-500-12      219792        205134        -6.67%
BenchmarkFieldSet/difference-500-12        778033        352579        -54.68%
BenchmarkFieldSet/insert-1000-12           5288591       2782317       -47.39%
BenchmarkFieldSet/has-1000-12              2800          2298          -17.93%
BenchmarkFieldSet/union-1000-12            2039674       692459        -66.05%
BenchmarkFieldSet/intersection-1000-12     436493        352916        -19.15%
BenchmarkFieldSet/difference-1000-12       1457956       738868        -49.32%

benchmark                                  old allocs     new allocs     delta
BenchmarkFieldSet/insert-20-12             227            130            -42.73%
BenchmarkFieldSet/has-20-12                7              6              -14.29%
BenchmarkFieldSet/union-20-12              129            84             -34.88%
BenchmarkFieldSet/intersection-20-12       43             60             +39.53%
BenchmarkFieldSet/difference-20-12         105            73             -30.48%
BenchmarkFieldSet/insert-50-12             729            435            -40.33%
BenchmarkFieldSet/has-50-12                8              7              -12.50%
BenchmarkFieldSet/union-50-12              424            190            -55.19%
BenchmarkFieldSet/intersection-50-12       89             117            +31.46%
BenchmarkFieldSet/difference-50-12         310            157            -49.35%
BenchmarkFieldSet/insert-100-12            2413           1309           -45.75%
BenchmarkFieldSet/has-100-12               10             9              -10.00%
BenchmarkFieldSet/union-100-12             1203           401            -66.67%
BenchmarkFieldSet/intersection-100-12      169            268            +58.58%
BenchmarkFieldSet/difference-100-12        801            357            -55.43%
BenchmarkFieldSet/insert-500-12            13179          7754           -41.16%
BenchmarkFieldSet/has-500-12               12             12             +0.00%
BenchmarkFieldSet/union-500-12             5862           2198           -62.50%
BenchmarkFieldSet/intersection-500-12      1013           1469           +45.01%
BenchmarkFieldSet/difference-500-12        4169           1886           -54.76%
BenchmarkFieldSet/insert-1000-12           29000          17208          -40.66%
BenchmarkFieldSet/has-1000-12              13             13             +0.00%
BenchmarkFieldSet/union-1000-12            11506          4118           -64.21%
BenchmarkFieldSet/intersection-1000-12     1841           2585           +40.41%
BenchmarkFieldSet/difference-1000-12       7994           3492           -56.32%
```

Even though intersection does more allocations, it's still somehow faster.