@lavalamp 2019-07-18
I made another change to permit it to cache the order for PathElement.Keys. It was apparently a huge deal.

```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkFieldSet/insert-20-12             54252         19315         -64.40%
BenchmarkFieldSet/has-20-12                1597          704           -55.92%
BenchmarkFieldSet/union-20-12              36040         11495         -68.10%
BenchmarkFieldSet/intersection-20-12       7631          5153          -32.47%
BenchmarkFieldSet/difference-20-12         25865         8763          -66.12%
BenchmarkFieldSet/insert-50-12             149605        62701         -58.09%
BenchmarkFieldSet/has-50-12                1776          873           -50.84%
BenchmarkFieldSet/union-50-12              95960         29569         -69.19%
BenchmarkFieldSet/intersection-50-12       16027         10213         -36.28%
BenchmarkFieldSet/difference-50-12         67187         20901         -68.89%
BenchmarkFieldSet/insert-100-12            498853        165516        -66.82%
BenchmarkFieldSet/has-100-12               2357          1287          -45.40%
BenchmarkFieldSet/union-100-12             246561        62777         -74.54%
BenchmarkFieldSet/intersection-100-12      34281         20617         -39.86%
BenchmarkFieldSet/difference-100-12        170985        46939         -72.55%
BenchmarkFieldSet/insert-500-12            2797068       770229        -72.46%
BenchmarkFieldSet/has-500-12               2607          1459          -44.04%
BenchmarkFieldSet/union-500-12             1095431       283355        -74.13%
BenchmarkFieldSet/intersection-500-12      219792        126864        -42.28%
BenchmarkFieldSet/difference-500-12        778033        260655        -66.50%
BenchmarkFieldSet/insert-1000-12           5288591       2120121       -59.91%
BenchmarkFieldSet/has-1000-12              2800          1360          -51.43%
BenchmarkFieldSet/union-1000-12            2039674       540864        -73.48%
BenchmarkFieldSet/intersection-1000-12     436493        213266        -51.14%
BenchmarkFieldSet/difference-1000-12       1457956       394617        -72.93%

benchmark                                  old allocs     new allocs     delta
BenchmarkFieldSet/insert-20-12             227            78             -65.64%
BenchmarkFieldSet/has-20-12                7              2              -71.43%
BenchmarkFieldSet/union-20-12              129            54             -58.14%
BenchmarkFieldSet/intersection-20-12       43             29             -32.56%
BenchmarkFieldSet/difference-20-12         105            43             -59.05%
BenchmarkFieldSet/insert-50-12             729            247            -66.12%
BenchmarkFieldSet/has-50-12                8              2              -75.00%
BenchmarkFieldSet/union-50-12              424            133            -68.63%
BenchmarkFieldSet/intersection-50-12       89             61             -31.46%
BenchmarkFieldSet/difference-50-12         310            101            -67.42%
BenchmarkFieldSet/insert-100-12            2413           813            -66.31%
BenchmarkFieldSet/has-100-12               10             3              -70.00%
BenchmarkFieldSet/union-100-12             1203           257            -78.64%
BenchmarkFieldSet/intersection-100-12      169            124            -26.63%
BenchmarkFieldSet/difference-100-12        801            214            -73.28%
BenchmarkFieldSet/insert-500-12            13179          4141           -68.58%
BenchmarkFieldSet/has-500-12               12             3              -75.00%
BenchmarkFieldSet/union-500-12             5862           1468           -74.96%
BenchmarkFieldSet/intersection-500-12      1013           739            -27.05%
BenchmarkFieldSet/difference-500-12        4169           1162           -72.13%
BenchmarkFieldSet/insert-1000-12           29000          8981           -69.03%
BenchmarkFieldSet/has-1000-12              13             3              -76.92%
BenchmarkFieldSet/union-1000-12            11506          2842           -75.30%
BenchmarkFieldSet/intersection-1000-12     1841           1330           -27.76%
BenchmarkFieldSet/difference-1000-12       7994           2238           -72.00%
```