@lavalamp 2019-07-18
I also looked at a mem & cpu profile while running the benchmarks and I don't think there's anything left that's easy to improve now. The latest change improved the original benchmarks, too:

```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkDeducedSimple-12                              422922        404477        -4.36%
BenchmarkDeducedNested-12                              1252984       1054765       -15.82%
BenchmarkDeducedNestedAcrossVersion-12                 1395454       1129654       -19.05%
BenchmarkLeafConflictAcrossVersion-12                  412367        393057        -4.68%
BenchmarkMultipleApplierRecursiveRealConversion-12     2481784       2242044       -9.66%

benchmark                                              old allocs     new allocs     delta
BenchmarkDeducedSimple-12                              1634           1525           -6.67%
BenchmarkDeducedNested-12                              4757           4180           -12.13%
BenchmarkDeducedNestedAcrossVersion-12                 5138           4524           -11.95%
BenchmarkLeafConflictAcrossVersion-12                  1672           1605           -4.01%
BenchmarkMultipleApplierRecursiveRealConversion-12     8154           7692           -5.67%
```

I'd like to squash and merge this, can someone review it first?